Change Dispensary.id from Integer to Long

+++ src/main/java/com/example/demo/application/services/DispensaryService.java
@@ -43,7 +43,7 @@ public interface DispensaryService {
       Integer adminId,
       Boolean enabled);
 
-  Optional<Dispensary> getById(Integer id);
+  Optional<Dispensary> getById(Long id);
 
-  Optional<Dispensary> deleteById(Integer id) throws NotFoundException;
+  Optional<Dispensary> deleteById(Long id) throws NotFoundException;
 }
