+++ src/main/java/com/example/demo/domain/repositories/DispensaryRepository.java
@@ -22,4 +22,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DispensaryRepository extends JpaRepository<Dispensary, Integer> {}
+public interface DispensaryRepository extends JpaRepository<Dispensary, Long> {}
